+++ docs/pages/components/loading/examples/ExProgrammatically.vue
@@ -1,15 +1,31 @@
 <template>
-    <button class="button is-primary is-medium"
-        @click="open">
-        Launch loading
-    </button>
+    <div>
+        <b-field>
+            <button class="button is-primary is-medium" @click="open">
+                Launch loading
+            </button>
+        </b-field>
+        <b-field>
+            <b-switch v-model="isFullPage">Display loader over full page</b-switch>
+        </b-field>
+        <b-notification ref="element" :closable="false">
+            Lorem ipsum dolor sit amet, consectetur adipiscing elit. Fusce id fermentum quam. Proin sagittis, nibh id hendrerit imperdiet, elit sapien laoreet elit
+        </b-notification>
+    </div>
 </template>
 
 <script>
     export default {
+        data() {
+            return {
+                isFullPage: true,
+            }
+        },
         methods: {
             open() {
-                const loadingComponent = this.$loading.open()
+                const loadingComponent = this.$loading.open({
+                    container: this.isFullPage ? null : this.$refs.element.$el
+                })
                 setTimeout(() => loadingComponent.close(), 3 * 1000)
             }
         }