+++ src/components/autocomplete/Autocomplete.spec.js
@@ -1,0 +1,10 @@
+import { shallow } from '@vue/test-utils'
+import BAutocomplete from '@components/autocomplete/Autocomplete'
+
+describe('BAutocomplete', () => {
+    it('is called', () => {
+        const wrapper = shallow(BAutocomplete)
+        expect(wrapper.name()).toBe('BAutocomplete')
+        expect(wrapper.isVueInstance()).toBeTruthy()
+    })
+})