+++ src/components/snackbar/Snackbar.spec.js
@@ -1,0 +1,14 @@
+import { shallow } from '@vue/test-utils'
+import BSnackbar from '@components/snackbar/Snackbar'
+
+describe('BSnackbar', () => {
+    HTMLElement.prototype.insertAdjacentElement = jest.fn()
+
+    it('is called', () => {
+        const wrapper = shallow(BSnackbar, {
+            attachToDocument: true
+        })
+        expect(wrapper.name()).toBe('BSnackbar')
+        expect(wrapper.isVueInstance()).toBeTruthy()
+    })
+})